@@ -2,20 +2,21 @@ package options
 
 import (
 	"context"
-	"log"
+	"log/slog"
 
 	"github.com/nats-io/nats.go"
 )
 
 type natsStub struct {
 	Verbose *bool
+	logger  *slog.Logger
 }
 
 func (s *natsStub) Subscribe(subj string, cb nats.MsgHandler) (*nats.Subscription, error) {
 	if !*s.Verbose {
 		return nil, nil
 	}
-	log.Println("NATS: subscribing to subj=", subj)
+	s.logger.Debug("NATS: subscribing to", "subj", subj)
 	return nil, nil
 }
 
@@ -23,7 +24,7 @@ func (s *natsStub) QueueSubscribe(subj, queue string, cb nats.MsgHandler) (*nats
 	if !*s.Verbose {
 		return nil, nil
 	}
-	log.Println("NATS: subscribing to subj=", subj, "queue=", queue)
+	s.logger.Debug("NATS: subscribing to", "subj", subj, "queue", queue)
 	return nil, nil
 }
 
@@ -31,7 +32,7 @@ func (s *natsStub) RequestMsgWithContext(ctx context.Context, m *nats.Msg) (*nat
 	if !*s.Verbose {
 		return m, nil
 	}
-	log.Println("NATS: req subj=", m.Subject, " len(data)=", len(m.Data), " data=", string(m.Data))
+	s.logger.Debug("NATS: req ", "subj", m.Subject, "len(data)", len(m.Data), "data", string(m.Data))
 	return m, nil
 }
 
@@ -39,7 +40,7 @@ func (s *natsStub) PublishMsg(m *nats.Msg) error {
 	if !*s.Verbose {
 		return nil
 	}
-	log.Println("NATS: pub subj=", m.Subject, " len(data)=", len(m.Data), " data=", string(m.Data))
+	s.logger.Debug("NATS: pub", "subj", m.Subject, "len(data)", len(m.Data), "data", string(m.Data))
 	return nil
 }
 
