@@ -3,14 +3,13 @@ package service
 import (
 	"context"
 	"crypto/sha256"
-	"encoding/base64"
 	"errors"
 	"fmt"
-	"log"
 	"slices"
 	"strings"
 	"time"
 
+	"github.com/cosmos/btcutil/base58"
 	"github.com/nats-io/nats.go"
 )
 
@@ -18,7 +17,7 @@ var ErrNotAvailable = fmt.Errorf("not available")
 
 func (b *Service) jsMakeHash(subjects ...string) string {
 	sum := sha256.Sum256([]byte(strings.Join(subjects, ",")))
-	return base64.RawStdEncoding.EncodeToString(sum[:])
+	return base58.Encode(sum[:])
 }
 
 func (b *Service) jsStreamName(hash string) string {
@@ -176,8 +175,9 @@ func (b *Service) attemptJSConsume(handler nats.MsgHandler, subject string) (*na
 	}
 
 	b.Group.Go(func() error {
-		log.Println("PullSubscribe loop start: ", subject)
-		defer log.Println("PullSubscribe loop exit: ", subject)
+		b.Logger.Info("PullSubscribe loop start", "subject", subject)
+		defer b.Logger.Info("PullSubscribe loop exit", "subject", subject)
+
 		for {
 			select {
 			case <-b.Context.Done():
@@ -197,7 +197,7 @@ func (b *Service) attemptJSConsume(handler nats.MsgHandler, subject string) (*na
 					continue
 				}
 				if errors.Is(err, nats.ErrBadSubscription) {
-					log.Printf("subscription to %s closed", subject)
+					b.Logger.Info("subscription closed", "subject", subject)
 					return nil
 				}
 				return fmt.Errorf("pulling message failed: %w", err)
@@ -205,7 +205,7 @@ func (b *Service) attemptJSConsume(handler nats.MsgHandler, subject string) (*na
 			for _, msg := range msgs {
 				handler(msg)
 				if err := msg.Ack(); err != nil {
-					log.Println("message ack failed: ", err)
+					b.Logger.Warn("message ack failed", err)
 				}
 			}
 		}
